% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_splice_exonic.R
\name{filter_splice_exonic}
\alias{filter_splice_exonic}
\title{Filter mySplice GRanges to retain only variants with at least one exon annotation in VEP}
\usage{
filter_splice_exonic(
  gr_splice,
  exon_keywords = c("exon", "exonic", "protein_coding", "missense", "nonsense")
)
}
\arguments{
\item{gr_splice}{A GRanges object returned by merge_splice_with_vep(), with a vep_info GRangesList column}

\item{exon_keywords}{A character vector of terms considered as exon annotations (default includes \"exon\")}
}
\value{
A filtered GRanges object
}
\description{
Filter mySplice GRanges to retain only variants with at least one exon annotation in VEP
}
