% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_read_names.R
\name{get_all_read_names}
\alias{get_all_read_names}
\title{Extract all read names from a mySplice GRanges object}
\usage{
get_all_read_names(gr)
}
\arguments{
\item{gr}{GRanges object with a 'supporting_reads' column}
}
\value{
A named list of character vectors, one per variant
}
\description{
Extract all read names from a mySplice GRanges object
}
