% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_names_for_row.R
\name{get_read_names_for_row}
\alias{get_read_names_for_row}
\title{Extract read names for a single variant row in mySplice GRanges}
\usage{
get_read_names_for_row(gr, row_index)
}
\arguments{
\item{gr}{GRanges object with a 'supporting_reads' column}

\item{row_index}{Integer index of the row (variant)}
}
\value{
A character vector of read names
}
\description{
Extract read names for a single variant row in mySplice GRanges
}
