% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_dependencies.R
\name{install_python_dependencies}
\alias{install_python_dependencies}
\title{Install Python dependencies for mySpliceR}
\usage{
install_python_dependencies(method = "auto")
}
\arguments{
\item{method}{Installation method ("auto", "virtualenv", or "conda")}
}
\description{
Install Python dependencies for mySpliceR
}
