% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_split_maf_files.R
\name{list_split_maf_files}
\alias{list_split_maf_files}
\title{List MAF split output files}
\usage{
list_split_maf_files(file_in, dir_out)
}
\arguments{
\item{file_in}{Original input MAF file.}

\item{dir_out}{Directory containing the split files.}
}
\value{
A character vector of full paths to the split files, sorted by suffix.
}
\description{
Given the original MAF filename and output directory,
this returns a sorted list of split MAF files generated by \code{split_maf()}.
}
