% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_splice_with_metadata_columns.R
\name{merge_splice_with_metadata_columns}
\alias{merge_splice_with_metadata_columns}
\title{Merge metadata columns from a GRanges to the GRanges from mySplice (1-to-1 match)}
\usage{
merge_splice_with_metadata_columns(gr_splice, gr_meta)
}
\arguments{
\item{gr_splice}{A GRanges object from mySplice(return_granges = TRUE)}

\item{gr_meta}{A GRanges with columns: sample, ref, alt, and metadata to attach}
}
\value{
A GRanges object with metadata columns from gr_meta added
}
\description{
Merge metadata columns from a GRanges to the GRanges from mySplice (1-to-1 match)
}
