% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_splice_with_vep.R
\name{merge_splice_with_vep}
\alias{merge_splice_with_vep}
\title{Merge mySplice output GRanges with VEP GRanges metadata}
\usage{
merge_splice_with_vep(gr_splice, gr_vep)
}
\arguments{
\item{gr_splice}{A GRanges object returned by mySplice(return_granges = TRUE)}

\item{gr_vep}{A VEP-annotated GRanges object with names like "chr1:26693521_C/T"}
}
\value{
A GRanges object with an added list-column \code{vep_info} containing matching VEP metadata rows as GRanges
}
\description{
Merge mySplice output GRanges with VEP GRanges metadata
}
