% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySplice.R
\name{mySplice}
\alias{mySplice}
\title{Detect mis-splicing events based on somatic mutations}
\usage{
mySplice(maf_file, bam_list, bed_file, output_file, min_reads = 5)
}
\arguments{
\item{maf_file}{Path to the input MAF file}

\item{bam_list}{Path to the TSV file listing sample ID and BAM paths}

\item{bed_file}{BED file with exon/intron junction annotations}

\item{output_file}{Path to output result file}

\item{min_reads}{Minimum number of supporting reads to report a novel junction (default 5)}
}
\description{
Detect mis-splicing events based on somatic mutations
}
