% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gtf_to_bed.R
\name{parse_gtf_to_bed}
\alias{parse_gtf_to_bed}
\title{Convert a Gencode GTF file to a BED-format exon junction DataFrame or file}
\usage{
parse_gtf_to_bed(gtf_path, output_bed_path = NULL)
}
\arguments{
\item{gtf_path}{Path to GTF file (required)}

\item{output_bed_path}{Optional path to save the BED file. If not provided, returns a data.frame}
}
\value{
A data.frame of BED-format exon junctions if \code{output_bed_path} is not given.
}
\description{
Convert a Gencode GTF file to a BED-format exon junction DataFrame or file
}
\examples{
df <- parse_gtf_to_bed("gencode.v36.annotation.gtf")
parse_gtf_to_bed("gencode.v36.annotation.gtf", "junctions.bed")
}
