% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_maf.R
\name{split_maf}
\alias{split_maf}
\title{Split a MAF file into smaller chunks}
\usage{
split_maf(file_in, dir_out, max_num)
}
\arguments{
\item{file_in}{Path to the input MAF file.}

\item{dir_out}{Directory where the split files will be written. It will be created if it doesn't exist.}

\item{max_num}{Maximum number of output files to create.}
}
\value{
Invisibly returns the system exit code (0 if successful).
}
\description{
This function splits a MAF (Mutation Annotation Format) file into approximately equal-sized chunks,
while preserving the header line in each output file. The splitting is handled by a Python script
bundled in the \code{inst/python} directory of the package.
}
\examples{
\dontrun{
  split_maf("inst/extdata/example.maf", "inst/extdata/splits", 5)
}
}
